It works

+++ respaldoProgramado/borrar/borrarRespaldoPlantilla.sh
@@ -18,11 +18,29 @@ while true; do
         # FIN RESPALDOS
         2>&1 >/dev/tty)
     echo $OPTIONS >tmp/oneline.tmp
+    eliminaciones=$(cat tmp/newline.tmp | wc -l)
+    if [[ $OPTIONS == "" ]]; then
+        clear
+        break
+    fi
+
     # tmp/oneline.tmp contiene los indices de los respaldos a borrar separados por espacios
     # Se crea un archivo llamado tmp/newline.tmp que contiene los indices de los respaldos a borrar separados por saltos de linea
     awk '{for (i=1; i<=NF; i++) print $i}' tmp/oneline.tmp >tmp/newline.tmp
     rm -f tmp/oneline.tmp
-
+    
+    for ((i=eliminaciones; i>=1; i--)); do
+        # Con sed se elimina la linea i de tmp/list.tmp
+        sed -i "$(<tmp/newline.tmp sed -n ${i}p)d" tmp/list.tmp
+    done
+    rm -f tmp/newline.tmp
+    
+    # re escribir crontab
+    head -$((numeroInicio)) /etc/crontab >tmp/crontab
+    cat tmp/list.tmp >>tmp/crontab
+    tail -n +$((numeroFin)) /etc/crontab >>tmp/crontab
+    rm -f tmp/list.tmp
+    mv tmp/crontab /etc/crontab
     break
     clear
 done
